Librarian Bot: Update dataset YAML metadata for model
This is a pull request to add a dataset, [`ctheodoris/Genecorpus-30M`](https://huggingface.co/datasets/ctheodoris/Genecorpus-30M), to the metadata for your model (defined in the `YAML` block of your model's `README.md`). 
The pull request was made by [librarian-bot](https://huggingface.co/librarian-bot) and used a combination of rules and/or machine learning to suggest this additional metadata.
If this suggestion is incorrect, feel free to close this pull request.  

Librarian Bot was made by [@davanstrien](https://huggingface.co/davanstrien); feel free to get in touch with feedback.