Upload 15 files
Deployable version of GeneFormer gene/cell classification and embedding extraction in a single function. Function parameters explained in the markdown file, example usage at the bottom of each python file. Let me know if anything is needed or if there are unresolved issues, and I can get to fixing them!